Protocol: realtime housekeeping
